replace old tests + ignore tune test for now + add happy-dom
